requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.7.0
numpy>=1.26.0
scikit-learn>=1.4.0
pydantic>=2.6.0
httpx>=0.27.0
pyyaml>=6.0
